Input fix

See merge request alialfred/FREDServer!69

+++ src/Parser/utility.cpp
@@ -24,11 +24,11 @@ vector<double> Utility::splitString2Num(const string &text, string by)
     {
         if (nums[j].find('x') != string::npos)
         {
-            result.push_back(double(stoi(nums[j].substr(nums[j].find('x') + 1), NULL, 16)));
+            result.push_back(double(stoll(nums[j].substr(nums[j].find('x') + 1), NULL, 16)));
         }
         else
         {
-            result.push_back(stof(nums[j]));
+            result.push_back(stod(nums[j]));
         }
     }
 
@@ -113,13 +113,13 @@ void Utility::printVector(const vector<int32_t> &data)
 
 double Utility::calculateEquation(string& equation, vector<string>& variables, vector<double>& values)
 {
-    map<string, int> varMap;
+    map<string, int64_t> varMap;
 
     for (size_t i = 0; i < values.size(); i++)
     {
         if (trunc(values[i]) == values[i])
         {
-            varMap[variables[i]] = int(values[i]);
+            varMap[variables[i]] = int64_t(values[i]);
         }
         else
         {
@@ -128,7 +128,7 @@ double Utility::calculateEquation(string& equation, vector<string>& variables, v
             {
                 equation.insert(pos + variables[i].size(), "/1000");
             }
-            varMap[variables[i]] = int(values[i] * 1000);
+            varMap[variables[i]] = int64_t(values[i] * 1000);
         }
     }
